Refactor ParkingSession endpoints and parameter types
Updated ParkingSessionController routes to nest under /api/[controller]/{lotId:long}/sessions and simplified endpoint paths. Changed session and lot ID parameters from int to long across controller, service interface, and implementation for consistency. Improved error handling and parameter validation in controller actions. Renamed PaymentRequired record property from Message to Reason for clarity.

+++ MobyPark/Services/Interfaces/IParkingSessionService.cs
@@ -26,6 +26,6 @@ public interface IParkingSessionService
     Task<StartSessionResult> StartSession(CreateParkingSessionDto sessionDto);
     Task<StartSessionResult> StartPaidSession(string licensePlate, long lotId, CreateCardInfoDto cardInfo);
     Task<StopSessionResult> StopSession(long id, StopParkingSessionDto dto);
-    Task<List<ParkingSessionModel>> GetAuthorizedSessionsAsync(long userId, int lotId, bool canManageSessions);
-    Task<GetSessionResult> GetAuthorizedSessionAsync(long userId, int lotId, int sessionId, bool canManageSessions);
+    Task<List<ParkingSessionModel>> GetAuthorizedSessionsAsync(long userId, long lotId, bool canManageSessions);
+    Task<GetSessionResult> GetAuthorizedSessionAsync(long userId, long lotId, long sessionId, bool canManageSessions);
 }
